hal: renesas: XTAL32M configuration bug fixes
This commit fixes issues related to XTAL32M configuration scheme which resulted in
not firing the XTAL32M Ready ISR and thus the device got stuck waiting for the
ISR to fire. With the fixes introduced the crystal is now configured to the normal
startup mode (32kHz) and its settling time is configured dynamically at each
XTAL32M Ready ISR firing so the optimum settling time is achieved.

Signed-off-by: Ioannis Karachalios <ioannis.karachalios.px@renesas.com>